+++ EldenRingSignatureResolver.cs
@@ -39,8 +39,8 @@ namespace DSDeaths {
             IntPtr process,
             IntPtr baseAddress,
             byte[] buffer,
-            int size,
-            ref int bytesRead);
+            UIntPtr size,
+            out UIntPtr bytesRead);
 
         public static bool TryResolve(
             IntPtr process,
@@ -147,14 +147,14 @@ namespace DSDeaths {
             }
 
             byte[] pointerStorage = new byte[sizeof(long)];
-            int bytesRead = 0;
+            UIntPtr bytesRead;
             if (!ReadProcessMemory(
                     process,
                     new IntPtr(match.PointerStorageAddress),
                     pointerStorage,
-                    pointerStorage.Length,
-                    ref bytesRead) ||
-                bytesRead != pointerStorage.Length) {
+                    new UIntPtr((uint)pointerStorage.Length),
+                    out bytesRead) ||
+                bytesRead.ToUInt64() != (ulong)pointerStorage.Length) {
                 int errorCode = Marshal.GetLastWin32Error();
                 error = "The resolved pointer storage could not be read: " + DescribeWin32Error(errorCode);
                 return false;
@@ -181,15 +181,15 @@ namespace DSDeaths {
                     uniqueEnd + EldenRingSignature.PatternLength - 1);
                 int readLength = checked((int)(readEnd - uniqueStart));
                 var buffer = new byte[readLength];
-                int bytesRead = 0;
+                UIntPtr bytesRead;
 
                 if (!ReadProcessMemory(
                         process,
                         new IntPtr(uniqueStart),
                         buffer,
-                        readLength,
-                        ref bytesRead) ||
-                    bytesRead != readLength) {
+                        new UIntPtr((uint)readLength),
+                        out bytesRead) ||
+                    bytesRead.ToUInt64() != (ulong)readLength) {
                     int errorCode = Marshal.GetLastWin32Error();
                     error = "Could not scan executable game memory at 0x" +
                             uniqueStart.ToString("X16") + ": " + DescribeWin32Error(errorCode);